[API-8172, API-8202] Adds `$promotions` to `$update_account` and moves `$iata_carrier_code`
