+++ src/main/java/com/siftscience/model/BaseAccountFieldSet.java
@@ -14,6 +14,7 @@ public abstract class BaseAccountFieldSet<T extends BaseAccountFieldSet<T>>
     @Expose @SerializedName("$payment_methods") private List<PaymentMethod> paymentMethods;
     @Expose @SerializedName("$billing_address") private Address billingAddress;
     @Expose @SerializedName("$shipping_address") private Address shippingAddress;
+    @Expose @SerializedName("$promotions") private List<Promotion> promotions;
     @Expose @SerializedName("$social_sign_on_type") private String socialSignOnType;
     @Expose @SerializedName("$merchant_profile") private MerchantProfile merchantProfile;
     @Expose @SerializedName(VERIFICATION_PHONE_NUMBER) private String verificationPhoneNumber;
@@ -105,6 +106,15 @@ public abstract class BaseAccountFieldSet<T extends BaseAccountFieldSet<T>>
 
     public T setVerificationPhoneNumber(String verificationPhoneNumber) {
         this.verificationPhoneNumber = verificationPhoneNumber;
+        return (T) this;
+    }
+
+    public List<Promotion> getPromotions() {
+        return promotions;
+    }
+
+    public T setPromotions(List<Promotion> promotions) {
+        this.promotions = promotions;
         return (T) this;
     }
 }
