+++ src/main/java/com/siftscience/model/Booking.java
@@ -14,7 +14,6 @@ public class Booking {
     @Expose @SerializedName("$currency_code") private String currencyCode;
     @Expose @SerializedName("$exchange_rate") private ExchangeRate exchangeRate;
     @Expose @SerializedName("$quantity") private Long quantity;
-    @Expose @SerializedName("$iata_carrier_code") private String iataCarrierCode;
     @Expose @SerializedName("$guests") private List<Guest> guests;
     @Expose @SerializedName("$segments") private List<Segment> segments;
     @Expose @SerializedName("$room_type") private String roomType;
@@ -93,15 +92,6 @@ public class Booking {
 
     public Booking setQuantity(Long quantity) {
         this.quantity = quantity;
-        return this;
-    }
-
-    public String getIataCarrierCode() {
-        return iataCarrierCode;
-    }
-
-    public Booking setIataCarrierCode(String iataCarrierCode) {
-        this.iataCarrierCode = iataCarrierCode;
         return this;
     }
