+++ src/main/java/com/siftscience/model/CreateAccountFieldSet.java
@@ -11,20 +11,10 @@ public class CreateAccountFieldSet extends BaseAccountFieldSet<CreateAccountFiel
     }
 
     @Expose @SerializedName("$account_types") private List<String> accountTypes;
-    @Expose @SerializedName("$promotions") private List<Promotion> promotions;
 
     @Override
     public String getEventType() {
         return "$create_account";
-    }
-
-    public List<Promotion> getPromotions() {
-        return promotions;
-    }
-
-    public CreateAccountFieldSet setPromotions(List<Promotion> promotions) {
-        this.promotions = promotions;
-        return this;
     }
 
     public List<String> getAccountTypes() { return accountTypes; }
