+++ src/main/java/com/siftscience/model/Segment.java
@@ -12,6 +12,7 @@ public class Segment {
     @Expose @SerializedName("$arrival_airport_code") private String arrivalAirportCode;
     @Expose @SerializedName("$departure_airport_code") private String departureAirportCode;
     @Expose @SerializedName("$fare_class") private String fareClass;
+    @Expose @SerializedName("$iata_carrier_code") private String iataCarrierCode;
 
 
     public Address getDepartureAddress() {
@@ -83,6 +84,23 @@ public class Segment {
 
     public Segment setFareClass(String fareClass) {
         this.fareClass = fareClass;
+        return this;
+    }
+
+    public String getIataCarrierCode() {
+        return iataCarrierCode;
+    }
+
+    /**
+     * Sets the IATA carrier code for the flight segment.
+     *
+     * <p>This field should only be set when the booking type is {@code $flight}.</p>
+     *
+     * @param iataCarrierCode The IATA carrier code.
+     * @return The updated {@code Segment} object.
+     */
+    public Segment setIataCarrierCode(String iataCarrierCode) {
+        this.iataCarrierCode = iataCarrierCode;
         return this;
     }
 }
